Tidying and test for escape_special_chars
