+++ preprocessors/mdbook-wikilinks/src/main.rs
@@ -1,7 +1,11 @@
-use mdbook_wikilinks::{handle_preprocessing, WikiLinks};
+use mdbook_wikilinks::WikiLinks;
 
 use clap::{App, Arg, SubCommand};
-use mdbook::errors::Error;
+use mdbook::{
+    preprocess::{CmdPreprocessor, Preprocessor},
+    errors::Error,
+};
+use std::io;
 
 pub fn make_app() -> App<'static, 'static> {
     App::new("wikilink")
@@ -21,4 +25,23 @@ fn main() -> Result<(), Error> {
     } else {
         handle_preprocessing(WikiLinks)
     }
+}
+
+pub fn handle_preprocessing(pre: impl Preprocessor) -> Result<(), Error> {
+    let (ctx, book) = CmdPreprocessor::parse_input(io::stdin())?;
+
+    if ctx.mdbook_version != mdbook::MDBOOK_VERSION {
+        eprintln!(
+            "Warning: The {} plugin was built against version {} of mdbook, \
+             but we're being called from version {}",
+            pre.name(),
+            mdbook::MDBOOK_VERSION,
+            ctx.mdbook_version
+        );
+    }
+
+    let processed_book = pre.run(&ctx, book)?;
+    serde_json::to_writer(io::stdout(), &processed_book)?;
+
+    Ok(())
 }
